requirements.txt
certifi==2017.4.17
chardet==3.0.3
idna==2.5
pkg-resources==0.0.0
PyYAML==3.12
requests==2.17.3
urllib3==1.21.1
